+++ axum-extra/src/extract/query.rs
@@ -1,6 +1,6 @@
-use axum::extract::FromRequestParts;
 use axum_core::__composite_rejection as composite_rejection;
 use axum_core::__define_rejection as define_rejection;
+use axum_core::extract::FromRequestParts;
 use http::{request::Parts, Uri};
 use serde_core::de::DeserializeOwned;